move set_epi64, setr_epi64 to x86
